+++ application/package.json
@@ -21,6 +21,7 @@
     "react": "^18.3.1",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.2",
     "styled-components": "^6.1.13",
     "vite-express": "^0.19.0"
   },
