+++ application/src/strategy/charts/BarChart.tsx
