+++ application/src/strategy/charts/LineChart.tsx
