+++ application/src/strategy/charts/MapChart.tsx
