+++ application/src/strategy/charts/PieChart.tsx
