+++ application/src/strategy/charts/TableChart.tsx
